+++ src/component/graphics/MovableBlockGraphicsComponent.cpp
@@ -10,8 +10,8 @@ void MovableBlockGraphicsComponent::draw(MovableBlock *block)
 	uint16_t blockX = block->getLocation()->getX();
 	uint16_t blockY = block->getLocation()->getY();
 
-	uint8_t displayBlockX = blockX - getDisplayStartX();
-	uint8_t displayBlockY = blockY - getDisplayStartY();
+	int16_t displayBlockX = blockX - getDisplayStartX();
+	int16_t displayBlockY = blockY - getDisplayStartY();
 
 	Sprites::drawOverwrite(MAP_CANVAS_X + displayBlockX, MAP_CANVAS_Y + displayBlockY, tileset, 14);
 }
